basics of dynamic memory allocation malloc/free
The malloc/free implementation remembers the size of each block as it is allocated, so it is not necessary to remind it of the size when freeing. Typically, the size is stored adjacent to the allocated block (header), which is why things usually break badly if the bounds of the allocated block are even slightly overstepped.